moved shortcuts to widgets
